tests/requirements.txt
pytest==7.4.3
pytest-bdd==7.0.1
docker==6.1.3
PyYAML==6.0.1
colorama==0.4.6
requests==2.31.0
tabulate==0.9.0
pytest-html==3.2.0
pytest-xdist==3.3.1